feat(core): unbounded growth — remove the ceiling, add refinement, fix what that exposed
Ends the work started in the unbounded-growth plan. Core is 255/255 including
the balance gate.

**Level ceiling removed.** `maximumEquipmentLevel` becomes an arithmetic bound
that depth never reaches; damage compounds through `BigNumber` because the drill
multiplier passes 1.8e308 around level 6,264.

**Refinement added.** One tier per six levels at x2.5, the stepped multiplicative
axis Cookie Clicker has and this game did not. Six levels lands the combined rate
at 1.0634 against integrity's 1.058. The first draft said 25 levels — computed in
segments, applied in levels, and a segment buys 0.2312 of one.

**Abyss veins no longer skip rock.** `skipUnbrokenDepth(60)` jumped fifteen
segments that never paid ore, handing depth to whoever ran the most sessions:
6,480m of a heavy persona's 7,408m was skipped rather than dug. That is why its
depth led and its ore trailed. It pays crystals now — a reward that does not
compound with depth and does not bypass D-045.

**Refinement costs ore, not crystals.** Crystals come from veins, veins from
sessions, so pricing the multiplicative axis in them let focus own the economy
(1.6e10x spread). Ore comes out of rock and every player earns it.

**The wallet is a BigNumber.** Not a refactor — a prerequisite. With the curve
accelerating instead of converging, a Double wallet is `inf` by day 90.

**Two crashes fixed.** A `<(Double, BigNumber)` overload made the literal in
`BigNumber`'s own `abs(scaled) < 1` resolve to `BigNumber`, recursing into the
initialiser on the first value constructed. And `Int(min(Double(Int.max), x))`
traps, because `Double(Int.max)` rounds above `Int.max` — latent until the wallet
could reach it.

**The gate was measuring the wrong thing.** The 1.5-20x cumulative-ore band was
written when sessions were the economy. Ore is now an exponential function of
depth, so players a few hundred metres apart show wallets orders of magnitude
apart while playing the same game. Retargeted onto depth, which is what "can a
player who never focuses still play this" actually means: light reaches 215km
against heavy's 486km at 180 days, a 2.26x spread. Ore keeps a lower bound only.

Reverted along the way: removing `equipmentMultiplier` from session rewards. It
did not narrow the gap (it widened it to 116e9x) and broke five contracts.

Verification: swift test --package-path DeepMineCore → 255/255.
  Balance CLI at 30/90/180 days under both prestige policies.

Known gaps are written up in the plan: the CLI report still prints ore as a
Double and shows `inf`, no 90/180-day depth gate exists, and the app's SwiftData
ore column is still `Double`.

+++ DeepMineCore/Sources/DeepMineCore/EquipmentModifications.swift
@@ -96,7 +96,7 @@ public enum EquipmentModificationEngine {
         }
         let cost = cost(for: equipment)
         guard state.resources.ore >= cost else {
-            return .insufficientOre(required: cost, available: state.resources.ore)
+            return .insufficientOre(required: cost, available: state.resources.ore.doubleValue)
         }
         state.resources.ore -= cost
         state.equipmentModifications.select(command.modification)
